+++ packages/ui-core/src/components/Select/Select.tsx
@@ -5,8 +5,13 @@ import SelectItem from './SelectItem';
 import * as equal from 'deep-equal';
 import cnCreate from 'utils/cn';
 import detectTouch from 'utils/detectTouch';
+import InputLabel from 'components/InputLabel/InputLabel';
 
 interface ISelectProps {
+    /** Field title */
+    label?: React.ReactNode;
+    /** Html id attribute */
+    id?: string;
     /** Header with the selected value */
     selectedTitle?: JSX.Element[] | Element[] | JSX.Element | string | Element;
     /** Selected value */
@@ -94,6 +99,8 @@ interface ISelectState {
 const cn = cnCreate('mfui-select');
 class Select extends React.Component<ISelectProps, ISelectState> {
     static propTypes = {
+        label: PropTypes.node,
+        id: PropTypes.string,
         selectedTitle: PropTypes.oneOfType([
             PropTypes.arrayOf(PropTypes.element),
             PropTypes.string,
@@ -356,7 +363,7 @@ class Select extends React.Component<ISelectProps, ISelectState> {
     }
 
     renderSearchField() {
-        const { inputPadding } = this.props;
+        const { inputPadding, id } = this.props;
 
         return (
             <input
@@ -372,6 +379,7 @@ class Select extends React.Component<ISelectProps, ISelectState> {
                 placeholder={this.props.placeholder}
                 maxLength={60}
                 autoComplete="off"
+                id={id}
             />
         );
     }
@@ -429,7 +437,7 @@ class Select extends React.Component<ISelectProps, ISelectState> {
             onChangeSearch, canOpen, className,
             name, icon, arrow, classNameControl,
             fontSize, fontColor, resultSize,
-            controlsPadding,
+            controlsPadding, label, id,
         } = this.props;
         const { focus, isOpen } = this.state;
 
@@ -451,6 +459,7 @@ class Select extends React.Component<ISelectProps, ISelectState> {
                 }, className)}
                 ref={this.getSelectNode}
             >
+                {label && <InputLabel htmlFor={id}>{label}</InputLabel>}
                 {!onChangeSearch && name && this.renderHiddenInput()}
                 <div
                     className={cn('control', { padding: controlsPadding }, classNameControl)}