+++ packages/ui-core/src/components/TextField/TextField.test.tsx
@@ -1,6 +1,12 @@
 import * as React from 'react';
-import { shallow, mount } from 'enzyme';
+import { shallow } from 'enzyme';
 import TextField from './TextField';
+import detectTouch from 'utils/detectTouch';
+import { ITextFieldProps } from './TextField';
+
+jest.mock('utils/detectTouch', () => ({
+    default: jest.fn().mockReturnValue(false),
+}));
 
 jest.mock('react-input-mask', () => {
     return {
@@ -8,72 +14,151 @@ jest.mock('react-input-mask', () => {
     };
 });
 
+const inputClassName = '.mfui-text-field__field';
+
+const props: ITextFieldProps = {
+    label: <div>label</div>,
+    color: 'default',
+    size: 'large',
+    noticeText: 'notice text',
+    commentText: 'comment text',
+    successText: 'success text',
+    autoFocus: true,
+    autocomplete: 'on',
+    disabled: true,
+    required: true,
+    type: 'text',
+    name: 'name',
+    placeholder: 'placeholder',
+    id: 'id',
+    value: 'value',
+    maxLength: 10,
+    defaultValue: 'default value',
+    customIcon: <div>custom icon</div>,
+    bigSpace: true,
+    className: 'class name',
+};
+
+function simulateEvent(eventName: string, eventType: string, params: {}) {
+    const data = 'data';
+    const mock = jest.fn();
+    const handler = { [eventName]: mock };
+    const wrapper = shallow(<TextField {...handler} {...params} />);
+
+    wrapper.find(inputClassName).simulate(eventType, data);
+
+    expect(mock).toHaveBeenCalledTimes(1);
+    expect(mock).toBeCalledWith(data);
+}
+
 describe('<TextField />', () => {
-    it('it renders TextField', () => {
+    beforeEach(() => {
+        jest.resetAllMocks();
+    });
+
+    it('renders without props', () => {
         const wrapper = shallow(<TextField />);
         expect(wrapper).toMatchSnapshot();
     });
 
-    it('it renders texts', () => {
-        const wrapper = shallow(<TextField commentText="text" successText="text" />);
+    it('renders with props', () => {
+        const wrapper = shallow(<TextField {...props} />);
         expect(wrapper).toMatchSnapshot();
     });
 
-    it('it renders without valid icon', () => {
-        const wrapper = shallow(<TextField valid={true} isHideIcon={true} />);
+    it('renders with mask', () => {
+        const wrapper = shallow(<TextField {...props} mask="(999) 999-99-99" maskChar=" " />);
         expect(wrapper).toMatchSnapshot();
     });
 
-    it('it renders without error icon', () => {
-        const wrapper = shallow(<TextField error={true} isHideIcon={true} />);
+    it('renders without no-touch mode', () => {
+        const typedDetectTouch = detectTouch as jest.Mock<() => boolean>;
+
+        typedDetectTouch.mockImplementation(() => true);
+
+        const wrapper = shallow(<TextField />);
+
         expect(wrapper).toMatchSnapshot();
     });
 
-    it('it renders mask input', () => {
-        const wrapper = shallow(<TextField mask="(999) 999-99-99" maskChar=" " />);
-        expect(wrapper).toMatchSnapshot();
+    describe('icon mod', () => {
+        it('renders with mod when customIcon passed', () => {
+            const wrapper = shallow(
+                <TextField isHideIcon={false} customIcon={<div>icon</div>} error={false} valid={false} />
+            );
+            expect(wrapper).toMatchSnapshot();
+        });
+
+        it('renders with mod when error is true', () => {
+            const wrapper = shallow(
+                <TextField isHideIcon={false} customIcon={undefined} error={true} valid={false} />
+            );
+            expect(wrapper).toMatchSnapshot();
+        });
+
+        it('renders with mod when valid is true', () => {
+            const wrapper = shallow(
+                <TextField isHideIcon={false} customIcon={undefined} error={false} valid={true} />
+            );
+            expect(wrapper).toMatchSnapshot();
+        });
+
+        it('renders without mod when isHideIcon is true', () => {
+            const wrapper = shallow(<TextField isHideIcon={true} />);
+            expect(wrapper).toMatchSnapshot();
+        });
     });
 
-    it('it renders error icon', () => {
-        const wrapper = shallow(<TextField error={true} noticeText="text" />);
-        expect(wrapper).toMatchSnapshot();
+    describe('status icons', () => {
+        it('renders when valid or error is true', () => {
+            const wrapper = shallow(
+                <TextField isHideIcon={false} customIcon={undefined} error={true} valid={true} />
+            );
+            expect(wrapper).toMatchSnapshot();
+        });
+
+        it('not renders when valid or error is false', () => {
+            const wrapper = shallow(
+                <TextField isHideIcon={false} customIcon={undefined} error={false} valid={false} />
+            );
+            expect(wrapper).toMatchSnapshot();
+        });
     });
 
-    it('it renders valid icon', () => {
-        const wrapper = shallow(<TextField valid={true} noticeText="text" />);
-        expect(wrapper).toMatchSnapshot();
+    describe('password type', () => {
+        it('renders with default input', () => {
+            const wrapper = shallow(<TextField type="password" />);
+            expect(wrapper).toMatchSnapshot();
+        });
+
+        it('renders with mask input', () => {
+            const wrapper = shallow(<TextField type="password" mask="(999) 999-99-99" maskChar=" " />);
+            expect(wrapper).toMatchSnapshot();
+        });
     });
 
-    it('it renders custom icon', () => {
-        const wrapper = shallow(<TextField customIcon={<img />} />);
-        expect(wrapper).toMatchSnapshot();
+    it('calls handlers on different inputs', () => {
+        const maskParams = { mask: '(999) 999-99-99', maskChar: ' ' };
+
+        simulateEvent('onChange', 'change', {});
+        simulateEvent('onChange', 'change', maskParams);
+
+        simulateEvent('onBlur', 'blur', {});
+        simulateEvent('onBlur', 'blur', maskParams);
+
+        simulateEvent('onFocus', 'focus', {});
+        simulateEvent('onFocus', 'focus', maskParams);
+
+        simulateEvent('onKeyUp', 'keyup', {});
+        simulateEvent('onKeyUp', 'keyup', maskParams);
     });
 
-    it('is checked shouldComponentUpdate', () => {
-        const wrapper = shallow(<TextField error={true} noticeText="text" />);
+    it('calls handleEyeToggle', () => {
+        const wrapper = shallow<TextField>(<TextField />);
 
-        wrapper.setProps({ error: true, noticeText: 'text' });
-        expect(wrapper).toMatchSnapshot();
-    });
+        wrapper.setState({ isPasswordHidden: false });
+        wrapper.instance().handleEyeToggle();
 
-    it('is checked shouldComponentUpdate with new props', () => {
-        const wrapper = shallow(<TextField error={true} noticeText="text" />);
-
-        wrapper.setProps({ error: false, noticeText: 'text' });
-        expect(wrapper).toMatchSnapshot();
-    });
-
-    it('is checked focus method', () => {
-        const wrapper = mount(<TextField />);
-        const instance = wrapper.instance() as TextField;
-
-        expect(() => instance.focus()).not.toThrow();
-    });
-
-    it('is checked blur method', () => {
-        const wrapper = mount(<TextField />);
-        const instance = wrapper.instance() as TextField;
-
-        expect(() => instance.blur()).not.toThrow();
+        expect(wrapper.state().isPasswordHidden).toBeTruthy();
     });
 });